Refactor Matty into package structure
