+++ matty/__main__.py
@@ -1,0 +1,6 @@
+"""Run Matty as `python -m matty`."""
+
+from .cli import app
+
+if __name__ == "__main__":
+    app()
